Refactor layer to handle minibatch at once (#181)
* get rid of unnecessary compiler warnings (C4018: '>=': signed/unsigned mismatch)

* refactor layer to handle a minibatch at once (previously individual input samples only)

* move new functions apply_cost_if_defined and add_sample_gradient to an anonymous namespace in order to hopefully make AppVeyor happy

* revert the changes to test_network.h

* minor type fixes to get rid of compiler warnings

* make quick changes to deconvolutional_layer so that the examples at least compile

* fix backward_activation to use all samples

* remove unused variables

* update picotest to fix operator precedence

* fix input data ordering

* fix overriding prev_delta

* change gradients to be per-sample as well

* remove unused variables

* fix indexing in convolutional_layer

* also in dropout_layer, have a different mask vector for each sample of a minibatch

* deconvolution_layer: fix allocating space for prev_out_ and cur_out_padded_

* add gradient check for minibatch

* minor: change types to fix some compiler warnings

+++ examples/mnist/test.cpp
@@ -44,7 +44,7 @@ double rescale(double x) {
 
 // convert tiny_cnn::image to cv::Mat and resize
 cv::Mat image2mat(image<>& img) {
-    cv::Mat ori(img.height(), img.width(), CV_8U, &img.at(0, 0));
+    cv::Mat ori(static_cast<int>(img.height()), static_cast<int>(img.width()), CV_8U, &img.at(0, 0));
     cv::Mat resized;
     cv::resize(ori, resized, cv::Size(), 3, 3, cv::INTER_AREA);
     return resized;
